Starts more accurately in sync without first seek
When starting the 2nd player, if we set rate at host time in the
completion of the seek to similar time of the 1st player, it seems that
often the synchronization is not good. Instead, if we set rate at host
time without seeking, it looks like we have perfect synchronization, as
determined by allowing both audio to be on for the same stream and not
hearing any dissonance (easier to hear than visual differences).

I feel like this makes sense... My impression of the set rate at host
time method was that it would sync the clocks of the two players and set
the playback to continue from the equivalent presentation time in the
new player as matched to the existing player based on the mapping we
obtained from EXT-X-PROGRAM-DATE-TIME. In this example, the presentation
timelines are the same, since I'm just using the same stream twice;
however, even if they weren't, the offset would be set by the PDT
mapping, and the rate of progress of decoded pictures will be driven by
the same clock, so we should be getting frame accurate synchronization
(assuming the two streams have frame accurate encoding sync).